cleanup and add linting
