+++ package.json
@@ -6,6 +6,8 @@
   "scripts": {
     "start": "webpack serve --open",
     "build": "webpack --color",
+    "fix": "yarn lint --fix",
+    "lint": "eslint --ext .js --ext .jsx ./*.js src/",
     "clean": "rimraf dist"
   },
   "author": "t.paine154@gmail.com",
@@ -22,12 +24,25 @@
   },
   "devDependencies": {
     "@babel/core": "^7.16.7",
+    "@babel/eslint-parser": "^7.16.5",
     "@babel/preset-env": "^7.16.7",
     "@babel/preset-react": "^7.16.7",
     "@finos/perspective-webpack-plugin": "^1.0.8",
     "@types/chroma-js": "^2.1.3",
     "babel-loader": "^8.2.3",
+    "eslint": "^8.6.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-import-resolver-node": "^0.3.6",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-json": "^3.1.0",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "html-webpack-plugin": "^5.3.2",
+    "prettier": "^2.5.1",
+    "prettier-eslint": "^13.0.0",
     "source-map-loader": "^3.0.0",
     "style-loader": "^3.3.0",
     "webpack": "^5.58.0",
