+++ webpack.config.js
@@ -12,38 +12,35 @@ const HtmlWebPackPlugin = require("html-webpack-plugin");
 const path = require("path");
 
 module.exports = {
-    mode: "development",
-    devtool: "source-map",
-    resolve: {
-        extensions: [".js", ".jsx"],
-    },
+  mode: "development",
+  devtool: "source-map",
+  resolve: {
+    extensions: [".js", ".jsx"],
+  },
 
-    plugins: [
-        new HtmlWebPackPlugin({
-            title: "Perspective React Example",
-            template: "./src/index.html",
-        }),
-        new PerspectivePlugin(),
+  plugins: [
+    new HtmlWebPackPlugin({
+      title: "Perspective React Example",
+      template: "./src/html/index.html",
+    }),
+    new PerspectivePlugin(),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.js(x?)$/,
+        enforce: "pre",
+        loader: "babel-loader",
+      },
+      {
+        test: /\.css$/,
+        exclude: /node_modules\/monaco-editor/,
+        use: [{loader: "style-loader"}, {loader: "css-loader"}],
+      },
     ],
-
-    module: {
-        rules: [
-            {
-                test: /\.js(x?)$/,
-                enforce: "pre",
-                loader: "babel-loader",
-            },
-            {
-                test: /\.css$/,
-                exclude: /node_modules\/monaco-editor/,
-                use: [{loader: "style-loader"}, {loader: "css-loader"}],
-            },
-        ],
-    },
-    devServer: {
-        static: [
-            {directory: path.join(__dirname, "dist")},
-            {directory: path.join(__dirname, "./static")},
-        ],
-    },
+  },
+  devServer: {
+    static: [{directory: path.join(__dirname, "dist")}, {directory: path.join(__dirname, "./static")}],
+  },
 };
